TGP-2125: Supplementary unit datatype inconsistency

+++ app/uk/gov/hmrc/tradergoodsprofilesdatastore/models/requests/UpdateRecordRequest.scala
@@ -29,7 +29,7 @@ case class UpdateRecordRequest(
   countryOfOrigin: Option[String] = None,
   category: Option[Int] = None,
   assessments: Option[Seq[Assessment]] = None,
-  supplementaryUnit: Option[Double] = None,
+  supplementaryUnit: Option[BigDecimal] = None,
   measurementUnit: Option[String] = None,
   comcodeEffectiveFromDate: Option[Instant] = None,
   comcodeEffectiveToDate: Option[Instant] = None
